+++ gama/postprocessing/ensemble.py
@@ -1,27 +1,33 @@
+import copy
 from collections import namedtuple
-import os
-import pickle
 import logging
+import random
 import time
 from typing import Optional, List
 
 import pandas as pd
+from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import OneHotEncoder
-import stopit
 
 from gama.genetic_programming.components import Individual
 from gama.postprocessing.base_post_processing import BasePostProcessing
+from gama.utilities.evaluation_library import EvaluationLibrary
 from gama.utilities.export import imports_and_steps_for_individual
 from gama.utilities.metrics import Metric, MetricType
-from gama.utilities.generic.async_evaluator import AsyncEvaluator
 
 log = logging.getLogger(__name__)
-Model = namedtuple("Model", ['name', 'individual', 'pipeline', 'predictions', 'validation_score'])
+Model = namedtuple("Model", ['name', 'individual', 'pipelines', 'predictions', 'validation_score'])
 
 
 class EnsemblePostProcessing(BasePostProcessing):
 
-    def __init__(self, time_fraction: float = 0.3, ensemble_size: Optional[int] = 25):
+    def __init__(
+            self,
+            time_fraction: float = 0.3,
+            ensemble_size: Optional[int] = 25,
+            hillclimb_size: Optional[int] = 10_000,
+            max_models: Optional[int] = 200
+    ):
         """ Ensemble construction per Caruana et al.
 
         Parameters
@@ -32,18 +38,26 @@ class EnsemblePostProcessing(BasePostProcessing):
             Total number of models in the ensemble.
             When a single model is chosen more than once, it will increase its weight in the ensemble and
             *does* count towards this maximum.
+        hillclimb_size: int, optional (default=10_000)
+            Number of predictions that are used to determine the ensemble score
+            during hillclimbing. If `None`, use all.
+        max_models: int, optional (default=200)
+            Only consider the best `max_models` number of models. If `None`, use all.
+            Consequently also sets the max number of unique models in the ensemble.
         """
         super().__init__(time_fraction)
         self._hyperparameters = dict(
             ensemble_size=(ensemble_size, 25),
             metric=(None, None),
-            cache=(None, None)
+            evaluation_library=(None, None),
+            hillclimb_size=(hillclimb_size, 10_000),
+            max_models=(max_models, 200)
         )
         self._ensemble = None
 
     def dynamic_defaults(self, gama: 'Gama'):
         self._overwrite_hyperparameter_default('metric', gama._metrics[0])
-        self._overwrite_hyperparameter_default('cache', gama._cache_dir)
+        self._overwrite_hyperparameter_default('evaluation_library', gama._evaluation_library)
 
     def post_process(self, x: pd.DataFrame, y: pd.Series, timeout: float, selection: List[Individual]) -> 'model':
         self._ensemble = build_fit_ensemble(
@@ -51,11 +65,11 @@ class EnsemblePostProcessing(BasePostProcessing):
             self.hyperparameters['ensemble_size'],
             timeout,
             self.hyperparameters['metric'],
-            self.hyperparameters['cache']
+            self.hyperparameters['evaluation_library']
         )
         return self._ensemble
 
-    def to_code(self) -> str:
+    def to_code(self, preprocessing: Optional[Pipeline] = None) -> str:
         voter = 'VotingClassifier' if isinstance(self._ensemble, EnsembleClassifier) else 'VotingRegressor'
         all_imports = {f"from sklearn.ensemble import {voter}", "from sklearn.pipeline import Pipeline"}
 
@@ -73,20 +87,36 @@ class EnsemblePostProcessing(BasePostProcessing):
 
         estimators = ','.join([f"('{i}', {name})" for i, name in enumerate(pipeline_names)])
         if isinstance(self._ensemble, EnsembleClassifier):
-            voting = ",'soft'" if self._ensemble._metric.requires_probabilities else 'hard'
+            voting = ",'soft'" if self._ensemble._metric.requires_probabilities else ", 'hard'"
         else:
             voting = ''  # This parameter does not exist for VotingRegressor
+
+        if preprocessing is not None:
+            # We don't want to export the mapping of categorical encoders
+            prepend = [(name, copy.copy(transformer))
+                       for name, transformer in preprocessing.steps]
+            for name, transformer in prepend:
+                transformer.mapping = None
+                all_imports = all_imports.union(
+                    {f"from {transformer.__module__} import {transformer.__class__.__name__}"}
+                )
+
         script = ('\n'.join(all_imports) + '\n\n' +
                   '\n\n'.join(pipeline_declarations) + '\n' +
                   f"ensemble = {voter}([{estimators}]{voting},{pipeline_weights})\n")
+        if preprocessing is not None:
+            steps = '\n,'.join([f"('{name}', {step})" for name, step in prepend])
+            script += f"pipeline = Pipeline([{steps},\n ('ensemble', ensemble)])\n"
         return script
 
 
 class Ensemble(object):
 
     def __init__(self, metric, y: pd.DataFrame,
-                 model_library=None, model_library_directory=None,
-                 shrink_on_pickle=True):
+                 evaluation_library: EvaluationLibrary = None,
+                 shrink_on_pickle=True,
+                 downsample_to: Optional[int] = 10_000,
+                 use_top_n_only: Optional[int] = 200):
         """
         Either model_library or model_library_directory must be specified.
         If model_library is specified, model_library_directory is ignored.
@@ -97,10 +127,8 @@ class Ensemble(object):
             Metric to optimize the ensemble towards.
         y: pandas.DataFrame
             True labels for the predictions made by the models in the library.
-        model_library:
+        evaluation_library: `gama.utilities.evaluation_library.EvaluationLibrary`
             A list of models from which an ensemble can be built.
-        model_library_directory: str
-            a directory containing results of model evaluations.
         shrink_on_pickle: bool (default=True)
             If True, remove memory-intensive attributes that are required before pickling.
             When unpickled, the model can be used to create predictions, but the ensemble can't be changed.
@@ -110,21 +138,35 @@ class Ensemble(object):
         elif not isinstance(metric, Metric):
             raise ValueError("metric must be specified as string or `gama.ea.metrics.Metric`.")
 
-        if model_library is None and model_library_directory is None:
-            raise ValueError("At least one of model_library or model_library_directory must be specified.")
-
-        if model_library is not None and model_library_directory is not None:
-            log.warning("model_library_directory will be ignored because model_library is also specified.")
+        if evaluation_library is None:
+            raise ValueError("`evaluation_library` is None but must be EvaluationLibrary.")
+        elif not isinstance(evaluation_library, EvaluationLibrary):
+            raise TypeError("`evaluation_library` must be of type gama.utilities.evaluation_library.EvaluationLibrary.")
 
         if not isinstance(y, (pd.Series, pd.DataFrame)):
             raise TypeError(f"`y_true` must be of type pandas.DataFrame or pandas.Series but is {type(y)}.")
 
         self._metric = metric
-        self._model_library_directory = model_library_directory
-        self._model_library = model_library if model_library is not None else []
+        self.evaluation_library = evaluation_library
+        self._model_library = []
+        self._use_top_n_only = use_top_n_only
         self._shrink_on_pickle = shrink_on_pickle
-        self._y = y
         self._prediction_transformation = None
+
+        if self.evaluation_library._sample is not None:
+            # If the library stores sampled predictions, we match that first.
+            y = y.iloc[self.evaluation_library._sample]
+
+        # Then apply even more sampling if requested.
+        if downsample_to is None or downsample_to >= len(y):
+            if downsample_to is not None:
+                log.info(f"Not downsampling because only {len(y)} samples were stored.")
+            self._y = y
+            self._prediction_sample = None
+        else:
+            log.info(f"Downsampling because training data exceeds {downsample_to} samples.")
+            self._prediction_sample = random.sample(range(len(y)), downsample_to)
+            self._y = y.iloc[self._prediction_sample]
 
         self._internal_score = None
         self._fit_models = None
@@ -135,9 +177,16 @@ class Ensemble(object):
     @property
     def model_library(self):
         if not self._model_library:
-            log.debug("Loading model library from disk.")
-            self._model_library = load_predictions(self._model_library_directory, self._prediction_transformation)
-            log.info("Loaded model library of size {} from disk.".format(len(self._model_library)))
+            self._model_library = []
+            for evaluation in self.evaluation_library.n_best(self._use_top_n_only):
+                if self._prediction_transformation:
+                    evaluation.predictions = self._prediction_transformation(evaluation.predictions)
+                if self._prediction_sample:
+                    evaluation.predictions = evaluation.predictions[self._prediction_sample]
+                self._model_library.append(
+                    Model(evaluation.individual.pipeline_str(), evaluation.individual, evaluation.estimators,
+                          evaluation.predictions, evaluation.score)
+                )
 
         return self._model_library
 
@@ -169,19 +218,16 @@ class Ensemble(object):
             log.warning("The ensemble already contained models. Overwriting the ensemble.")
             self._models = {}
 
-        sorted_ensembles = reversed(sorted(self.model_library, key=lambda m: m.validation_score))
-
         # Since the model library only features unique models, we do not need to check for duplicates here.
-        selected_models = list(sorted_ensembles)[:n]
-        for model in selected_models:
+        for model in self.model_library[:n]:
             self._add_model(model)
 
         log.debug("Initial ensemble created with score {}".format(self._ensemble_validation_score()))
 
     def _add_model(self, model, add_weight=1):
         """ Adds a specific model to the ensemble or increases its weight if it already is contained. """
-        model, weight = self._models.pop(model.pipeline, (model, 0))
-        self._models[model.pipeline] = (model, weight + add_weight)
+        model, weight = self._models.pop(model.individual._id, (model, 0))
+        self._models[model.individual._id] = (model, weight + add_weight)
         log.debug("Assigned a weight of {} to model {}".format(weight + add_weight, model.name))
 
     def expand_ensemble(self, n: int):
@@ -231,20 +277,25 @@ class Ensemble(object):
         if timeout <= 0:
             raise ValueError("timeout must be greater than 0.")
 
-        self._fit_models = []
-        futures = set()
-        with stopit.ThreadingTimeout(timeout) as c_mgr, AsyncEvaluator() as async_:
-            for (model, weight) in self._models.values():
-                futures.add(async_.submit(fit_and_weight, (model.pipeline, X, y, weight)))
+        self._fit_models = [(estimator, weight)
+                            for (model, weight) in self._models.values()
+                            for estimator in model.pipelines]
+        # for (model, weight) in self._models.values():
 
-            for _ in self._models.values():
-                future = async_.wait_next()
-                pipeline, weight = future.result
-                if weight > 0:
-                    self._fit_models.append((pipeline, weight))
-
-        if not c_mgr:
-            log.info("Fitting of ensemble stopped early.")
+        # self._fit_models = []
+        # futures = set()
+        # with stopit.ThreadingTimeout(timeout) as c_mgr, AsyncEvaluator() as async_:
+        #     for (model, weight) in self._models.values():
+        #         futures.add(async_.submit(fit_and_weight, (model.pipeline, X, y, weight)))
+        #
+        #     for _ in self._models.values():
+        #         future = async_.wait_next()
+        #         pipeline, weight = future.result
+        #         if weight > 0:
+        #             self._fit_models.append((pipeline, weight))
+        #
+        # if not c_mgr:
+        #     log.info("Fitting of ensemble stopped early.")
 
     def _get_weighted_mean_predictions(self, X, predict_method='predict'):
         weighted_predictions = []
@@ -276,23 +327,6 @@ class Ensemble(object):
             # alternatively, one could just save the number of classes instead.
 
         return self.__dict__.copy()
-
-
-def load_predictions(cache_dir, prediction_transformation=None):
-    models = []
-    for file in os.listdir(cache_dir):
-        if file.endswith('.pkl'):
-            file_name = os.path.join(cache_dir, file)
-            if os.stat(file_name).st_size > 0:
-                # We check file size, because writing to disk may be interrupted if the process was terminated due
-                # to a restart/timeout. I can not find specifications saying that any interrupt of pickle.dump leads
-                # to 0-sized files, but in practice this seems to case so far.
-                with open(os.path.join(cache_dir, file), 'rb') as fh:
-                    individual, predictions, scores = pickle.load(fh)
-                if prediction_transformation:
-                    predictions = prediction_transformation(predictions)
-                models.append(Model(individual.pipeline_str(), individual, individual.pipeline, predictions, scores))
-    return models
 
 
 def fit_and_weight(args):
@@ -332,7 +366,9 @@ class EnsembleClassifier(Ensemble):
         self._one_hot_encoder = OneHotEncoder(categories='auto').fit(y_as_squeezed_array)
 
         if self._metric.requires_probabilities:
-            self._y = self._one_hot_encoder.transform(y_as_squeezed_array).toarray()
+            self._y = self._one_hot_encoder.transform(self._y.values.reshape(-1, 1)).toarray()
+            if self._prediction_sample is not None:
+                self._y = self._y[self._prediction_sample]
         else:
             def one_hot_encode_predictions(predictions):
                 return self._one_hot_encoder.transform(predictions.reshape(-1, 1))
@@ -383,17 +419,17 @@ class EnsembleRegressor(Ensemble):
         return self._get_weighted_mean_predictions(X)
 
 
-def build_fit_ensemble(x, y, ensemble_size: int, timeout: float,
-                       metric: Metric, cache: str, encoder: Optional[object]=None) -> Ensemble:
+def build_fit_ensemble(x, y, ensemble_size: int, timeout: float, metric: Metric,
+                       evaluation_library: EvaluationLibrary, encoder: Optional[object] = None) -> Ensemble:
     """ Construct an Ensemble of models from cache, optimizing for metric and fit to (x, y). """
     start_build = time.time()
 
     log.debug('Building ensemble.')
     if metric.task_type == MetricType.CLASSIFICATION:
-        ensemble = EnsembleClassifier(metric, y, model_library_directory=cache)
+        ensemble = EnsembleClassifier(metric, y, evaluation_library=evaluation_library)
         ensemble._label_encoder = encoder
     elif metric.task_type == MetricType.REGRESSION:
-        ensemble = EnsembleRegressor(metric, y, model_library_directory=cache)
+        ensemble = EnsembleRegressor(metric, y, evaluation_library=evaluation_library)
     else:
         raise ValueError(f"Unknown metric task type {metric.task_type}")