+++ docs/source/releases.rst
@@ -1,6 +1,24 @@
 Release Notes
 =============
 
+Version 20.1.0
+--------------
+Features:
+ - Encoding of ARFF files may now be specified with the `encoding` parameter in {fit/predict/score}_arff calls.
+ - Set `max_pipeline_length` on initialization to impose a maximum number of steps in your pipelines.
+
+Bugfixes:
+ - Reading ARFF markers (such as @data and @attribute) is now correctly case insensitive.
+
+Maintenance:
+ - Evaluation results are no longer saved to disk but kept in memory only.
+   Consequently the `cache_dir` hyperparameter has been removed.
+
+Changes:
+ - Pipelines fitted during search are now used in the ensemble, instead of retraining the pipeline.
+ - Ordinal Encoding and One Hot Encoding are now applied outside of 5-fold CV.
+   This is for computational reasons, as all levels of a categorical variable are known this shouldn't make a difference.
+
 Version 20.0.0
 --------------
 Features: